Fixed RADI launch hierarchy
